#82 feat: create board with auth
