+++ src/lib/api/board.ts
@@ -1,5 +1,8 @@
+/* eslint-disable import/no-cycle */
+
 import { Board, CreateBoardPayload } from '@/types'
-import { get, post } from './base'
+import { get } from './base'
+import { authFetch } from './customFetch/authFetch'
 
 export const getBoard = async (id: string): Promise<Board> => {
   const res = await get(`/api/v1/boards/${id}`, {
@@ -14,7 +17,8 @@ export const getBoard = async (id: string): Promise<Board> => {
 export const postBoard = async (
   payload: CreateBoardPayload,
 ): Promise<string> => {
-  const res = await post('/api/v1/boards', {
+  const res = await authFetch('/api/v1/boards', {
+    method: 'POST',
     body: JSON.stringify(payload),
   })
