+++ src/lib/api/user.ts
@@ -1,6 +1,8 @@
+/* eslint-disable import/no-cycle */
+
 import { put } from '@/lib/api/base'
 import { WithdrawUserPayload } from '@/types'
-// import { authFetch } from './customFetch/authFetch'
+import { authFetch } from './customFetch/authFetch'
 
 export const withdraw = async (body: WithdrawUserPayload) => {
   return put('/api/v1/user/withdraw', {
@@ -8,19 +10,10 @@ export const withdraw = async (body: WithdrawUserPayload) => {
   })
 }
 
-export const changeNickname = async (nickName: string, token: string) => {
-  return put('/api/v1/user/nickname', {
-    headers: {
-      Authorization: `Bearer ${token}`,
+export const viewProfile = async () => {
+  return authFetch('/api/v1/user/profile', {
+    next: {
+      tags: ['profile'],
     },
-    body: JSON.stringify({ nickName }),
   })
 }
-
-// export const viewProfile = async () => {
-//   return authFetch('/api/v1/user/profile', {
-//     next: {
-//       tags: ['profile'],
-//     },
-//   })
-// }
